+++ cplusplus/BufferedOutput.cpp
@@ -10,7 +10,7 @@
  * governing permissions and limitations under the License.
  */
 
-#include "BufferedOutput.h"
+#include "zdw/BufferedOutput.h"
 
 
 namespace adobe {
