+++ cplusplus/status_output.cpp
@@ -10,7 +10,7 @@
  * governing permissions and limitations under the License.
  */
 
-#include "status_output.h"
+#include "zdw/status_output.h"
 #include <stdarg.h>
 #include <stdio.h>
