+++ cplusplus/zdw/BufferedInput.h
