+++ cplusplus/zdw/BufferedOutput.h
