+++ cplusplus/zdw/UnconvertFromZDW.h
@@ -13,11 +13,10 @@
 #ifndef UNCONVERTFROMZDW_H
 #define UNCONVERTFROMZDW_H
 
-#include "includes.h"
-
-#include "BufferedInput.h"
-#include "BufferedOutput.h"
-#include "status_output.h"
+#include "zdw/includes.h"
+#include "zdw/BufferedInput.h"
+#include "zdw/BufferedOutput.h"
+#include "zdw/status_output.h"
 
 #include <map>
 #include <string>
